* commit '2402d770c2b9791f4faec29a1c8214debc6f63ac':
  Merge pull request #2244 in HDFFV/hdf5 from ~KMU/hdf5:bugfix/intel_warnings to develop
  Merge pull request #2071 in HDFFV/hdf5 from ~KMU/hdf5:bugfix/intel_warnings to develop
  Merge pull request #2234 in HDFFV/hdf5 from ~KMU/hdf5:bugfix/intel_warnings to develop

+++ src/H5Oattr.c
@@ -315,7 +315,7 @@ H5O_attr_encode(H5F_t *f, uint8_t *p, const void *mesg)
 
     /* The character encoding for the attribute's name, in later versions */
     if(attr->shared->version >= H5O_ATTR_VERSION_3)
-        *p++ = attr->shared->encoding;
+        *p++ = (uint8_t)attr->shared->encoding;
 
     /* Write the name including null terminator */
     HDmemcpy(p, attr->shared->name, name_len);
