+++ src/H5SMmessage.c
@@ -301,7 +301,7 @@ H5SM_message_encode(uint8_t *raw, const void *_nrecord, void *_ctx)
     /* Sanity check */
     HDassert(ctx);
 
-    *raw++ = message->location;
+    *raw++ = (uint8_t)message->location;
     UINT32ENCODE(raw, message->hash);
 
     if(message->location == H5SM_IN_HEAP) {
